Fix in incorrect or non-portable format specifiers
Use the specifiers for fixed-bit integers contained in `<cinttypes>`
macros. These are portable across platforms with different definitions
for int, long int, etc., and our previous specifier usage was not
consistent, anyway. Use `%zu` for `size_t`, as it is specified to be
used. Also fix one signed specifier that should have been unsigned.

Related to #350; I wanted to update the code for printing floats, but
immediately found incorrect printing for other number types.

+++ Core/KernelSML/src/sml_AgentSML.cpp
@@ -38,6 +38,7 @@
 #endif
 
 #include <cassert>
+#include <cinttypes>
 
 using namespace sml ;
 
@@ -1184,7 +1185,7 @@ bool AgentSML::RemoveInputWME(int64_t clientTimeTag)
     //if (kDebugInput)
     //{
     //    std::string printInput1 = this->ExecuteCommandLine("print --internal --depth 2 I2") ;
-    //  PrintDebugFormat("%s\nLooking for %ld", printInput1.c_str(), timetag) ;
+    //  PrintDebugFormat("%s\nLooking for %" SCNu64, printInput1.c_str(), timetag) ;
     //}
 
     // The wme is already gone so no work to do
@@ -1580,7 +1581,7 @@ void AgentSML::ReplayInputWMEs()
         {
             // add-wme
             char timetagString[25];
-            SNPRINTF(timetagString, 25, "%ld", static_cast<long int>(ca.clientTimeTag));
+            SNPRINTF(timetagString, 25, "%" SCNd64, ca.clientTimeTag);
 
             if (!AddInputWME(ca.Add()->id.c_str(), ca.Add()->attr.c_str(), ca.Add()->value.c_str(), ca.Add()->type, timetagString))
             {
